Extract FastAPI-Users config from main.py
